feat(core): add table groups_resources_automatic

+++ perun-db/postgres.sql
@@ -1,4 +1,4 @@
--- database version 3.1.85 (don't forget to update insert statement at the end of file)
+-- database version 3.1.86 (don't forget to update insert statement at the end of file)
 
 -- VOS - virtual organizations
 create table vos (
@@ -763,8 +763,25 @@ create table groups_resources (
   constraint grres_res_fk foreign key (resource_id) references resources(id)
 );
 
+-- GROUPS_RESOURCES_AUTOMATIC - groups automatically assigned to resource through source group
+create table groups_resources_automatic (
+	group_id integer not null,     --identifier of group (groups.id)
+	resource_id integer not null,  --identifier of resource (resources.id)
+	source_group_id integer not null,  --identifier of source group (groups.id)
+	created_at timestamp default statement_timestamp() not null,
+	created_by varchar default user not null,
+	modified_at timestamp default statement_timestamp() not null,
+	modified_by varchar default user not null,
+	created_by_uid integer,
+	modified_by_uid integer,
+	constraint grresaut_grp_res_sgrp_u unique (group_id,resource_id,source_group_id),
+	constraint grresaut_gr_fk foreign key (group_id) references groups(id),
+	constraint grresaut_res_fk foreign key (resource_id) references resources(id),
+	constraint grresaut_sgr_fk foreign key (source_group_id) references groups(id)
+);
+
 create function relation_group_resource_exist(integer, integer) returns integer as
-    'select count(1) from groups_resources where group_id=$1 and resource_id=$2;' language sql;
+    'select count(1) from (SELECT group_id, resource_id FROM groups_resources UNION SELECT group_id, resource_id FROM groups_resources_automatic) gr_res where group_id=$1 and resource_id=$2;' language sql;
 
 create table groups_resources_state (
 	group_id integer not null,
@@ -779,13 +796,18 @@ create table groups_resources_state (
 
 create function delete_group_resource_status() returns trigger as
 	'
-    begin
-    	delete from groups_resources_state where group_id=OLD.group_id and resource_id=OLD.resource_id;
-		return OLD;
-   	end;
+	begin
+		if relation_group_resource_exist(OLD.group_id, OLD.resource_id) = 0 then
+			delete from groups_resources_state where group_id=OLD.group_id and resource_id=OLD.resource_id;
+		end if;
+	return OLD;
+	end;
 	' language plpgsql;
 
 create trigger after_deleting_from_groups_resources after delete on groups_resources
+	for each row execute procedure delete_group_resource_status();
+
+create trigger after_deleting_from_groups_resources_automatic after delete on groups_resources_automatic
 	for each row execute procedure delete_group_resource_status();
 
 -- MEMBER_ATTR_VALUES - values of attributes assigned to members
@@ -1610,6 +1632,9 @@ create index idx_fk_grres_gr on groups_resources(group_id);
 create index idx_fk_grres_res on groups_resources(resource_id);
 create index idx_fk_grres_s_gr on groups_resources_state(group_id);
 create index idx_fk_grres_s_res on groups_resources_state(resource_id);
+create index idx_fk_grres_a_gr on groups_resources_automatic(group_id);
+create index idx_fk_grres_a_res on groups_resources_automatic(resource_id);
+create index idx_fk_grres_a_sgr on groups_resources_automatic(source_group_id);
 create index idx_fk_grpmem_gr on groups_members(group_id);
 create index idx_fk_grpmem_mem on groups_members(member_id);
 create index idx_fk_grpmem_memtype on groups_members(membership_type);
@@ -1744,6 +1769,7 @@ grant all on roles to perun;
 grant all on authz to perun;
 grant all on groups_resources to perun;
 grant all on groups_resources_state to perun;
+grant all on groups_resources_automatic to perun;
 grant all on groups_members to perun;
 grant all on application_mails to perun;
 grant all on application_mail_texts to perun;
@@ -1779,7 +1805,7 @@ grant all on members_sponsored to perun;
 grant all on groups_to_register to perun;
 
 -- set initial Perun DB version
-insert into configurations values ('DATABASE VERSION','3.1.85');
+insert into configurations values ('DATABASE VERSION','3.1.86');
 
 -- insert membership types
 insert into membership_types (id, membership_type, description) values (1, 'DIRECT', 'Member is directly added into group');
